fix another race
